+++ firmware/friend_bot/ota_service.h
@@ -6,6 +6,8 @@ class OtaService {
   void update();
 
  private:
+  char ssid_[33] = {};
+  char password_[65] = {};
   bool wifiAttempted_ = false;
   bool wifiConnected_ = false;
   unsigned long lastWifiRetryMs_ = 0;